third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/background-size-expected.txt

This is a testharness.js-based test.
[FAIL] Can set 'background-size' to a length: -3.14em
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
[FAIL] Can set 'background-size' to a percent: -3.14%
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
Harness: the test ran to completion.
